fix a libzipfile bug
The bug is that when find the file in zip it only compare filename
prefix instead of filename.

For example:
we want to get "platform.img.gz", however, there is no
platform.img.gz in zip but platform.img is there.
we'll got platform.img by mistake.

This patch fix this error.

Signed-off-by: Chengwei Yang <chengwei.yang@intel.com>