Build fastboot for windows with libz static linked
On Windows, libz may not installed. It's fine if none of libz installed
because the windows launcher will complain libz1.dll not found. So the
user can search and install it.

However, If there is a libz 64bit installed, it will link to it and
complains

	"The application was unable to start correctly (0xc000007b). Click
	OK to close the application."

This is quite confusing.

Compare with distributing another file, I'd like to enlarge the fastboot
binary.

Change-Id: I5e9af52d322692b60627f535865dc04778e9620f
Signed-off-by: Chengwei Yang <chengwei.yang@intel.com>